Style Changes
Mainly changed the format as per the last meeting with the group
leaders. Modified the layout and linked to WOM.

I am going to switch focus to the discourse community at this point. I
am not getting much in the way of content from the group so I will wait
on that.